+++ packages/adapter-fauna/test/index.ts
@@ -1,4 +1,0 @@
-import { testAdapter } from "@lucia-auth/adapter-test";
-import { db, adapter } from "./db.js";
-
-testAdapter(adapter, db);