+++ packages/adapter-prisma/test/db.ts
@@ -1,45 +1,51 @@
-import { Database } from "@lucia-auth/adapter-test";
 import { PrismaClient } from "@prisma/client";
 import prisma from "../src/index.js";
 import { convertSession } from "../src/utils.js";
 import { KeySchema, LuciaError } from "lucia-auth";
+import type { LuciaQueryHandler } from "@lucia-auth/adapter-test";
 
 const client = new PrismaClient();
 export const adapter = prisma(client)(LuciaError);
 
-export const db: Database = {
-	getUsers: async () => {
-		return await client.user.findMany();
+export const db: LuciaQueryHandler = {
+	user: {
+		get: async () => {
+			return await client.user.findMany();
+		},
+		insert: async (user) => {
+			await client.user.create({
+				data: user
+			});
+		},
+		clear: async () => {
+			await client.user.deleteMany();
+		}
 	},
-	getSessions: async () => {
-		const sessions = await client.session.findMany();
-		return sessions.map((session) => convertSession(session));
+	session: {
+		get: async () => {
+			const sessions = await client.session.findMany();
+			return sessions.map((session) => convertSession(session));
+		},
+		insert: async (session) => {
+			await client.session.create({
+				data: session
+			});
+		},
+		clear: async () => {
+			await client.session.deleteMany();
+		}
 	},
-	getKeys: async () => {
-		return (await client.key.findMany()) as unknown as KeySchema[];
-	},
-	insertUser: async (user) => {
-		await client.user.create({
-			data: user
-		});
-	},
-	insertSession: async (session) => {
-		await client.session.create({
-			data: session
-		});
-	},
-	insertKey: async (key) => {
-		await client.key.create({
-			data: key
-		});
-	},
-	clearUsers: async () => {
-		await client.user.deleteMany();
-	},
-	clearSessions: async () => {
-		await client.session.deleteMany();
-	},
-	clearKeys: async () => {
-		await client.key.deleteMany();
+	key: {
+		get: async () => {
+			return (await client.key.findMany()) as unknown as KeySchema[];
+		},
+		insert: async (key) => {
+			await client.key.create({
+				data: key
+			});
+		},
+		clear: async () => {
+			await client.key.deleteMany();
+		}
 	}
 };