+++ packages/adapter-test/src/types.ts
@@ -1,5 +1,0 @@
-import { UserSchema } from "lucia-auth"
-
-export type UserSchemaWithAttributes = UserSchema & {
-	username: string
-}